+++ Assets/uNode3/Core/CodeGenerator/CodeGeneratorNode.cs
@@ -44,10 +44,10 @@ namespace MaxyGames {
 					}
 				}
 				if(port.UseDefaultValue) {
-					if(port.IsOptional && (port.defaultValue == null || port.defaultValue.targetType == MemberData.TargetType.None)) {
+					if(port.IsOptional && (port.DefaultValue == null || port.DefaultValue.targetType == MemberData.TargetType.None)) {
 						return Value(port.OptionalValue, setVariable: setVariable, autoConvert: autoConvert);
 					}
-					return Value(port.defaultValue, setVariable: setVariable, autoConvert: autoConvert);
+					return Value(port.DefaultValue, setVariable: setVariable, autoConvert: autoConvert);
 				}
 				else {
 					if(!port.isConnected) {